+++ pyppeteer_ghost_cursor/pyppeteer/__init__.py
@@ -1,0 +1,3 @@
+from pyppeteer_ghost_cursor import _logger
+from .spoof import create_cursor
+from .mouse_helper import install_mouse_helper
