+++ pyppeteer_ghost_cursor/pyppeteer/mouse_helper.py
@@ -1,0 +1,9 @@
+from pyppeteer.page import Page
+from typing import Coroutine
+from pathlib import Path
+
+
+async def install_mouse_helper(page: Page) -> Coroutine[None, None, None]:
+    await page.evaluateOnNewDocument(
+        Path(__file__).parent.joinpath("../js/mouseHelper.js").read_text()
+    )
